serving/api/requirements.txt
fastapi
uvicorn
pydantic
pillow
numpy
pandas
python-multipart
ultralytics